pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "org-bot"
version = "0.1.0"
description = "org-mode bot for Telegram"
readme = "README.md"
authors = [
    { name = "Georgy Grigoryev", email = "iamkarlson@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "functions-framework==3.4.0",
    "Flask~=2.3.2",
    "python-telegram-bot==22.3",
    "PyGithub==1.59.1",
    "defopt",
    "sentry-sdk==2.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
